server: move httpTestServer out of testserver.go
Release note: None
