+++ pkg/server/testserver.go
@@ -13,16 +13,12 @@ package server
 import (
 	"bytes"
 	"context"
-	"encoding/base64"
 	"fmt"
 	"net"
 	"net/http"
-	"net/http/cookiejar"
-	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
-	"sync"
 	"time"
 
 	"github.com/cenkalti/backoff"
@@ -43,7 +39,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/security"
-	"github.com/cockroachdb/cockroach/pkg/server/serverpb"
 	"github.com/cockroachdb/cockroach/pkg/server/status"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/sql"
@@ -59,7 +54,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/metric"
 	addrutil "github.com/cockroachdb/cockroach/pkg/util/netutil/addr"
-	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
 	"github.com/cockroachdb/cockroach/pkg/util/retry"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	"github.com/cockroachdb/cockroach/pkg/util/tracing"
@@ -307,23 +301,6 @@ type TestServer struct {
 	*Server
 	// httpTestServer provides the HTTP APIs of TestTenantInterface.
 	*httpTestServer
-}
-
-type httpTestServer struct {
-	t struct {
-		// We need a sub-struct to avoid ambiguous overlap with the fields
-		// of *Server, which are also embedded in TestServer.
-		authentication *authenticationServer
-		sqlServer      *SQLServer
-	}
-	// authClient is an http.Client that has been authenticated to access the
-	// Admin UI.
-	authClient [2]struct {
-		httpClient http.Client
-		cookie     *serverpb.SessionCookie
-		once       sync.Once
-		err        error
-	}
 }
 
 var _ serverutils.TestServerInterface = &TestServer{}
@@ -841,16 +818,6 @@ func (ts *TestServer) WriteSummaries() error {
 	return ts.node.writeNodeStatus(context.TODO(), time.Hour, false)
 }
 
-// AdminURL implements TestServerInterface.
-func (ts *httpTestServer) AdminURL() string {
-	return ts.t.sqlServer.execCfg.RPCContext.Config.AdminURL().String()
-}
-
-// GetHTTPClient implements TestServerInterface.
-func (ts *httpTestServer) GetHTTPClient() (http.Client, error) {
-	return ts.t.sqlServer.execCfg.RPCContext.GetHTTPClient()
-}
-
 // UpdateChecker implements TestServerInterface.
 func (ts *TestServer) UpdateChecker() interface{} {
 	return ts.Server.updates
@@ -873,22 +840,6 @@ func authenticatedUserNameNoAdmin() security.SQLUsername {
 	return security.MakeSQLUsernameFromPreNormalizedString(authenticatedUserNoAdmin)
 }
 
-// GetAdminAuthenticatedHTTPClient implements the TestServerInterface.
-func (ts *httpTestServer) GetAdminAuthenticatedHTTPClient() (http.Client, error) {
-	httpClient, _, err := ts.getAuthenticatedHTTPClientAndCookie(authenticatedUserName(), true)
-	return httpClient, err
-}
-
-// GetAuthenticatedHTTPClient implements the TestServerInterface.
-func (ts *httpTestServer) GetAuthenticatedHTTPClient(isAdmin bool) (http.Client, error) {
-	authUser := authenticatedUserName()
-	if !isAdmin {
-		authUser = authenticatedUserNameNoAdmin()
-	}
-	httpClient, _, err := ts.getAuthenticatedHTTPClientAndCookie(authUser, isAdmin)
-	return httpClient, err
-}
-
 type v2AuthDecorator struct {
 	http.RoundTripper
 
@@ -898,88 +849,6 @@ type v2AuthDecorator struct {
 func (v *v2AuthDecorator) RoundTrip(r *http.Request) (*http.Response, error) {
 	r.Header.Add(apiV2AuthHeader, v.session)
 	return v.RoundTripper.RoundTrip(r)
-}
-
-func (ts *httpTestServer) getAuthenticatedHTTPClientAndCookie(
-	authUser security.SQLUsername, isAdmin bool,
-) (http.Client, *serverpb.SessionCookie, error) {
-	authIdx := 0
-	if isAdmin {
-		authIdx = 1
-	}
-	authClient := &ts.authClient[authIdx]
-	authClient.once.Do(func() {
-		// Create an authentication session for an arbitrary admin user.
-		authClient.err = func() error {
-			// The user needs to exist as the admin endpoints will check its role.
-			if err := ts.createAuthUser(authUser, isAdmin); err != nil {
-				return err
-			}
-
-			id, secret, err := ts.t.authentication.newAuthSession(context.TODO(), authUser)
-			if err != nil {
-				return err
-			}
-			rawCookie := &serverpb.SessionCookie{
-				ID:     id,
-				Secret: secret,
-			}
-			// Encode a session cookie and store it in a cookie jar.
-			cookie, err := EncodeSessionCookie(rawCookie, false /* forHTTPSOnly */)
-			if err != nil {
-				return err
-			}
-			cookieJar, err := cookiejar.New(nil)
-			if err != nil {
-				return err
-			}
-			url, err := url.Parse(ts.t.sqlServer.execCfg.RPCContext.Config.AdminURL().String())
-			if err != nil {
-				return err
-			}
-			cookieJar.SetCookies(url, []*http.Cookie{cookie})
-			// Create an httpClient and attach the cookie jar to the client.
-			authClient.httpClient, err = ts.t.sqlServer.execCfg.RPCContext.GetHTTPClient()
-			if err != nil {
-				return err
-			}
-			rawCookieBytes, err := protoutil.Marshal(rawCookie)
-			if err != nil {
-				return err
-			}
-			authClient.httpClient.Transport = &v2AuthDecorator{
-				RoundTripper: authClient.httpClient.Transport,
-				session:      base64.StdEncoding.EncodeToString(rawCookieBytes),
-			}
-			authClient.httpClient.Jar = cookieJar
-			authClient.cookie = rawCookie
-			return nil
-		}()
-	})
-
-	return authClient.httpClient, authClient.cookie, authClient.err
-}
-
-func (ts *httpTestServer) createAuthUser(userName security.SQLUsername, isAdmin bool) error {
-	if _, err := ts.t.sqlServer.internalExecutor.ExecEx(context.TODO(),
-		"create-auth-user", nil,
-		sessiondata.InternalExecutorOverride{User: security.RootUserName()},
-		fmt.Sprintf("CREATE USER %s", userName.Normalized()),
-	); err != nil {
-		return err
-	}
-	if isAdmin {
-		// We can't use the GRANT statement here because we don't want
-		// to rely on CCL code.
-		if _, err := ts.t.sqlServer.internalExecutor.ExecEx(context.TODO(),
-			"grant-admin", nil,
-			sessiondata.InternalExecutorOverride{User: security.RootUserName()},
-			"INSERT INTO system.role_members (role, member, \"isAdmin\") VALUES ('admin', $1, true)", userName.Normalized(),
-		); err != nil {
-			return err
-		}
-	}
-	return nil
 }
 
 // MustGetSQLCounter implements TestServerInterface.
